feat: Add tests for IPv6 and PortScan stages, including filtering and delta port logic
